# db: make tests faster