The runtime for the top-level package tests has been creeping up. This is an assorted set of small adjustments to improve the running time.

#### db: make TestCompaction/l0_to_lbase_compaction faster

This test took ~1s, now it takes 0.06s.

#### db: make TestCompaction/value_separation faster

This test was taking ~1s, now it's under 0.1s.

#### db: reduce TestOpenCloseOpenClose configurations

This saves about 0.5s.

#### db: remove duplicate TestFileCacheClockPro

The genericcache code already has this test.

#### db: reduce numOps in TestWALFailoverRandomized

This reduces run time from 4s to ~1.5s on average.

#### db: reduce numOps in TestWALHardCrashRandomized

This reduces the average run time from ~4s to ~0.5s.