@RaduBerinde 2025-09-18
Hm, `TestCompaction/value_separation` is now flaky..  The reason is just file numbering. Seems like the creation of a new WAL can happen during the flush and that affects the table numbers. How do other compaction tests go around this issue? I guess previously we had bigger files so an sstable wouldn't finish so quickly.. but relying on that seems fragile.

```
    logger.go:15: [JOB 4] flushing 2 memtables (1.3MB) to L0
    logger.go:15: [JOB 4] flushing: sstable created 000005
    logger.go:15: [JOB 4] flushing: blob file created 000007
    logger.go:15: [JOB 5] WAL created 000006
```
vs
```
    logger.go:15: [JOB 4] WAL created 000005
    logger.go:15: [JOB 5] flushing 3 memtables (1.3MB) to L0
    logger.go:15: [JOB 5] flushing: sstable created 000006
    logger.go:15: [JOB 5] flushing: blob file created 000007
```

I guess the previous